GH-20: Bring `getDefaultISO2Code` and `parsePhoneNumber` back
